add Entities, DTOs, Repo, Service

+++ src/main/java/pl/com/mike/developer/PremiseGetResponse.java
@@ -8,7 +8,7 @@ public class PremiseGetResponse {
     private Integer number;
     private Integer floor;
     private Double surfacePerSqMeter;
-    private Double pricePerSqMeter;
+    private BigDecimal pricePerSqMeter;
     private BigDecimal totalPrice;
     private Integer numberOfRooms;
     private String technicalStatus;
@@ -18,12 +18,12 @@ public class PremiseGetResponse {
     private Boolean isGarden;
     private Boolean isTerrace;
     private Boolean isLoggia;
-    private Integer buildingId;
+    private Long buildingId;
 
     public PremiseGetResponse() {
     }
 
-    public PremiseGetResponse(Long id, String type, Integer number, Integer floor, Double surfacePerSqMeter, Double pricePerSqMeter, BigDecimal totalPrice, Integer numberOfRooms, String technicalStatus, String salesStatus, String exposure, Boolean isBalcony, Boolean isGarden, Boolean isTerrace, Boolean isLoggia, Integer buildingId) {
+    public PremiseGetResponse(Long id, String type, Integer number, Integer floor, Double surfacePerSqMeter, BigDecimal pricePerSqMeter, BigDecimal totalPrice, Integer numberOfRooms, String technicalStatus, String salesStatus, String exposure, Boolean isBalcony, Boolean isGarden, Boolean isTerrace, Boolean isLoggia, Long buildingId) {
         this.id = id;
         this.type = type;
         this.number = number;
@@ -82,11 +82,11 @@ public class PremiseGetResponse {
         this.surfacePerSqMeter = surfacePerSqMeter;
     }
 
-    public Double getPricePerSqMeter() {
+    public BigDecimal getPricePerSqMeter() {
         return pricePerSqMeter;
     }
 
-    public void setPricePerSqMeter(Double pricePerSqMeter) {
+    public void setPricePerSqMeter(BigDecimal pricePerSqMeter) {
         this.pricePerSqMeter = pricePerSqMeter;
     }
 
@@ -162,11 +162,11 @@ public class PremiseGetResponse {
         isLoggia = loggia;
     }
 
-    public Integer getBuildingId() {
+    public Long getBuildingId() {
         return buildingId;
     }
 
-    public void setBuildingId(Integer buildingId) {
+    public void setBuildingId(Long buildingId) {
         this.buildingId = buildingId;
     }
 }
